Pause Menu Integration
I created a pause menu.
Added an Input action for pausing.
In the player script i added a section that, when the player does the action to pause the game, determines whether the game is currently paused or not.  I
f the game is not paused, the pause menu should appear. If not, the pause menu should be deleted.